Task_3_CLI/requirements.txt
pandas>=1.5.0
python-dotenv>=1.0.0